requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
redis>=5.0,<6.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
Pillow>=10.0,<11.0
django-htmx>=1.17,<2.0
